First refactor for Linux support.
* Unified directory layout (lowercase directories with UpperCamelCase file names)
