+++ src/IOpSys.h
@@ -1,6 +1,6 @@
 #pragma  once
 
-#include "Systems/KeyboardInputSystem.h"
+#include "systems/KeyboardInputSystem.h"
 
 class IOpSys {
 public:
